network and api/token checks in journal

+++ app/src/main/java/com/gitfit/android/ui/home/journal/JournalViewModel.kt
@@ -6,6 +6,8 @@ import com.gitfit.android.data.local.db.entity.Activity
 import com.gitfit.android.data.local.prefs.PreferenceProvider
 import com.gitfit.android.data.local.prefs.User
 import com.gitfit.android.data.remote.ResultWrapper
+import com.gitfit.android.data.remote.ResultWrapper.NetworkConnectivityError
+import com.gitfit.android.data.remote.ResultWrapper.Success
 import com.gitfit.android.repos.ActivityRepository
 import com.gitfit.android.repos.ActivityTypeRepository
 import com.gitfit.android.repos.GitFitAPIRepository
@@ -39,33 +41,43 @@ class JournalViewModel(
         setLoading(false)
     }
 
-    private suspend fun loadActivityTypes() = withContext(Dispatchers.IO) {
-        val activityTypes = gitFitAPIRepository.getActivityTypes()
+    private suspend fun loadActivityTypes() {
+        when(val response = gitFitAPIRepository.getActivityTypes()) {
+            is NetworkConnectivityError -> withContext(Dispatchers.Main) {
+                navigator()?.showToast("No network connection.")
+            }
+            is Success -> {
+                val activityTypes = response.value
 
-        activityTypes?.let {
-            activityTypeRepository.deleteAll()
-            val stringActivityTypes = activityTypes.mapTo(HashSet()) { it.name }
-            if (stringActivityTypes == ACTIVITY_TYPES) {
-                activityTypeRepository.insertList(activityTypes)
-            } else {
-                activityTypeRepository.insertList(activityTypes.filter { ACTIVITY_TYPES.contains(it.name) })
-                info("Server activity types don't match implemented ones. Make sure you have implemented all activity types!")
+                activityTypeRepository.deleteAll()
+                val stringActivityTypes = activityTypes.mapTo(HashSet()) { it.name }
+                if (stringActivityTypes == ACTIVITY_TYPES) {
+                    activityTypeRepository.insertList(activityTypes)
+                } else {
+                    activityTypeRepository.insertList(activityTypes.filter { ACTIVITY_TYPES.contains(it.name) })
+                    info("Server activity types don't match implemented ones. Make sure you have implemented all activity types!")
+                }
+            }
+            else -> withContext(Dispatchers.Main) {
+                navigator()?.showToast("Server connection error.")
             }
         }
     }
 
     private suspend fun loadActivities(user: User) = withContext(Dispatchers.IO) {
         when(val response = gitFitAPIRepository.getActivities(user.username, user.token)) {
-            is ResultWrapper.NetworkConnectivityError -> println('A')
-            is ResultWrapper.GenericError -> println('B')
-            is ResultWrapper.NetworkError -> println('C')
-            is ResultWrapper.Success -> {
+            is NetworkConnectivityError -> withContext(Dispatchers.Main) {
+                navigator()?.showToast("No network connection.")
+            }
+            is Success -> {
                 val activities = response.value
                     .filter { ACTIVITY_TYPES.contains(it.type) }
                     .map { a -> Activity.fromActivityResponse(a) }
 
                 activityRepository.deleteAll()
                 activityRepository.insertList(activities)
+            } else -> withContext(Dispatchers.Main) {
+                navigator()?.showToast("Server connection error.")
             }
         }
     }
